requirements.txt
jupyterlab
pandas
matplotlib
numpy
scikit-learn
sklearn-genetic
skrebate
